Move fetched data processing functions into separate file
